Add files via upload
This code is to estimate the simplification error. It’s an an extension to ngsolve / netgen to estimate the effect of model simplifications on a quantity of interest aimed at electrostatics problems.

Development of this code has been supported by the Marie Curie Initial training Network INSIST. Funding is provided by the EU Seventh Framework Programme (FP7) under grant agreement n° 289361.